test(select): Move portal test story
