+++ modules/_labs/select/react/stories/stories.tsx
@@ -12,8 +12,6 @@ import {
 } from '@workday/canvas-system-icons-web';
 import {colors, typeColors} from '@workday/canvas-kit-react-core';
 import FormField from '../../../../form-field/react/index';
-import {Button} from '../../../../button/react';
-import Modal from '../../../../modal/react';
 import Select from '../lib/Select';
 import {RenderOptionFunction} from '../lib/SelectBase';
 import README from '../README.md';
@@ -21,7 +19,7 @@ import README from '../README.md';
 const hintText = 'Helpful text goes here.';
 const hintId = 'error-desc-id';
 
-const options = [
+export const options = [
   {label: 'E-mail', value: 'email'},
   {label: 'Phone', value: 'phone'},
   {label: 'Fax (disabled)', value: 'fax', disabled: true},
@@ -33,9 +31,9 @@ const options = [
   },
 ];
 
-const simpleOptions = ['California', 'Florida', 'New York', 'Pennsylvania', 'Texas'];
+export const simpleOptions = ['California', 'Florida', 'New York', 'Pennsylvania', 'Texas'];
 
-const manyOptions = [
+export const manyOptions = [
   {label: 'Atlanta', value: 'atlanta'},
   {label: 'Austin', value: 'austin'},
   {label: 'Beaverton', value: 'beaverton'},
@@ -50,14 +48,14 @@ const manyOptions = [
   {label: 'San Mateo', value: 'san-mateo'},
 ];
 
-const customOptions = [
+export const customOptions = [
   {value: 'Activity Stream', data: {icon: activityStreamIcon}},
   {value: 'Avatar', data: {icon: avatarIcon}},
   {value: 'Upload Cloud', data: {icon: uploadCloudIcon}},
   {value: 'User', data: {icon: userIcon}},
 ];
 
-const customRenderOption: RenderOptionFunction = option => {
+export const customRenderOption: RenderOptionFunction = option => {
   const iconColor = option.focused ? typeColors.inverse : colors.blackPepper100;
   return (
     <div style={{alignItems: 'center', display: 'flex', padding: '3px 0'}}>
@@ -66,141 +64,6 @@ const customRenderOption: RenderOptionFunction = option => {
     </div>
   );
 };
-
-const SelectModal = () => {
-  const [open, setOpen] = React.useState(false);
-  const buttonRef = React.useRef<HTMLButtonElement>() as React.RefObject<HTMLButtonElement>; // cast to keep buttonRef happy
-  const openModal = () => {
-    setOpen(true);
-  };
-  const closeModal = () => {
-    setOpen(false);
-    if (buttonRef.current) {
-      buttonRef.current.focus();
-    }
-  };
-
-  return (
-    <div>
-      <Button buttonRef={buttonRef} onClick={openModal} variant={Button.Variant.Primary}>
-        Show Modal
-      </Button>
-
-      <Modal heading="Modal with Select" open={open} handleClose={closeModal}>
-        <p>The menu for this Select should break out of the Modal.</p>
-        <FormField label="Label" inputId="select-modal">
-          {controlComponent(<Select name="city" options={manyOptions} />)}
-        </FormField>
-        <Button style={{marginRight: '16px'}} onClick={closeModal} variant={Button.Variant.Primary}>
-          Submit
-        </Button>
-        <Button onClick={closeModal} variant={Button.Variant.Secondary}>
-          Cancel
-        </Button>
-      </Modal>
-    </div>
-  );
-};
-
-const SandboxContainer = ({children, style = {}, ...elemProps}) => {
-  return (
-    <div
-      style={{
-        border: `1px solid ${colors.soap600}`,
-        marginTop: 50,
-        overflow: 'hidden',
-        padding: 20,
-        ...style,
-      }}
-      {...elemProps}
-    >
-      {children}
-    </div>
-  );
-};
-
-storiesOf('Labs|Select/React', module)
-  .addParameters({component: Select})
-  .addDecorator(withReadme(README))
-  .add('Sandbox', () => (
-    <div>
-      <h3>Sandbox</h3>
-      <SandboxContainer>
-        <p>
-          All bordered containers have overflow set to hidden. Menus are rendered using portals and
-          should break out of their containers.
-        </p>
-        <FormField label="Label" inputId="select-default-top-label">
-          {controlComponent(<Select name="contact" options={options} />)}
-        </FormField>
-        <p>
-          Menus for Select with grow set to true should resize automatically as the Select grows.
-          Activate this menu and resize the window to see the menu grow to match the width of the
-          Select.
-        </p>
-        <FormField label="Label (Grow)" inputId="select-default-top-label-grow" grow={true}>
-          {controlComponent(<Select name="city" options={manyOptions} grow={true} />)}
-        </FormField>
-      </SandboxContainer>
-      <SandboxContainer>
-        <p>
-          Menus should flip upwards automatically if there isn't enough space in the viewport for
-          them to extend downwards.
-        </p>
-        <FormField
-          label="Label"
-          inputId="select-alert-top-label"
-          error={FormField.ErrorType.Alert}
-          hintText={hintText}
-        >
-          {controlComponent(<Select name="contact" options={options} />)}
-        </FormField>
-        <FormField
-          label="Label (Grow)"
-          inputId="select-alert-top-label-grow"
-          error={FormField.ErrorType.Alert}
-          grow={true}
-          hintText={hintText}
-        >
-          {controlComponent(<Select grow={true} name="city" options={manyOptions} />)}
-        </FormField>
-      </SandboxContainer>
-      <SandboxContainer>
-        <p>Menus should behave the same with left-labeled FormFields.</p>
-        <FormField
-          label="Label"
-          labelPosition={FormField.LabelPosition.Left}
-          inputId="select-default-left-label"
-        >
-          {controlComponent(<Select name="contact" options={options} />)}
-        </FormField>
-        <FormField
-          label="Label (Grow)"
-          labelPosition={FormField.LabelPosition.Left}
-          inputId="select-default-left-label-grow"
-          grow={true}
-        >
-          {controlComponent(<Select name="city" options={manyOptions} grow={true} />)}
-        </FormField>
-      </SandboxContainer>
-      <SandboxContainer>
-        <p>Menus should break out of Modals.</p>
-        <SelectModal />
-      </SandboxContainer>
-      <SandboxContainer>
-        <p>This Select is forced to display its menu upwards since it's at the bottom the page.</p>
-        <FormField
-          label="Label"
-          inputId="select-last"
-          error={FormField.ErrorType.Error}
-          hintText={hintText}
-          hintId={hintId}
-        >
-          {controlComponent(<Select name="contact" options={options} />)}
-        </FormField>
-      </SandboxContainer>
-    </div>
-  ));
 
 storiesOf('Labs|Select/React/Top Label', module)
   .addParameters({component: Select})
